Make some style changes
