+++ src/styles/styles.scss
@@ -109,3 +109,8 @@ footer {
   text-align: center;
   font-size: 1.5rem;
 }
+
+.greeting {
+  margin: 0 40px;
+  font-size: 1.2rem;
+}
